Document capturing {{ matrix_domain }}/_matrix with Traefik
When running this playbook I discovered that my server could not federate correctly, and was failing the Matrix federation checker due to invalid SSL certs.
I discovered this was because I am hosting matrix.domain.com on a different server/ip than I am hosting domain.com, and that the federation sends requests to the IP/domain specified by the SRV record, but containing the host header for the raw domain.com.

To resolve this, Traefik must have a rule to capture domain.com/_matrix traffic and send it to Synapse, but it also must have a valid cert for the raw domain.com. This would not work if using a standard .well-known challenge, however if you are able to use a DNS challenge it seems to be fine.

I have documented the problem and my simple configuration change here in case it is valuable to someone, however I would like to make clear that this **might** be just an ugly hack, and if there is a cleaner way to handle this problem we should document that instead.